requirements.txt
Flask

Jinja2
Werkzeug

itsdangerous
gunicorn

openai==1.14.2
python-dotenv==1.0.1
pillow==10.2.0

